+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.3.1
+- Add `--list`
+  - Will list all the collections and episodes selected and quit
+- Alias `--download-all` to `-a`
+- Alias `--episodes` to `-e`
+- Fix a problem with single episodes and `--episodes`
+
 ## 1.3.0
 - `--episodes`
 - `--dont-autoselect-quality`